# #1101 Use `computeOnSegments` for casting strings to numeric