--- a/src/CastMsg.chpl
+++ b/src/CastMsg.chpl
@@ -9,7 +9,7 @@ module CastMsg {
   use SysError;
   use ServerErrorStrings;
   use ServerConfig;
-  use CommAggregation;
+  use Cast;
 
   private config const logLevel = ServerConfig.logLevel;
   const castLogger = new Logger(logLevel);
@@ -124,21 +124,22 @@ module CastMsg {
       }
       when "str" {
           const strings = getSegString(name, st);
+          const errors = opt.toLower() : ErrorMode;
           select targetDtype {
               when "int64" {
-                  return new MsgTuple(castStringToSymEntry(strings, st, int), MsgType.NORMAL);
+                return new MsgTuple(castStringToSymEntry(strings, st, int, errors), MsgType.NORMAL);
               }
               when "uint8" {
-                  return new MsgTuple(castStringToSymEntry(strings, st, uint(8)), MsgType.NORMAL);
+                  return new MsgTuple(castStringToSymEntry(strings, st, uint(8), errors), MsgType.NORMAL);
               }
               when "uint64" {
-                  return new MsgTuple(castStringToSymEntry(strings, st, uint), MsgType.NORMAL);
+                  return new MsgTuple(castStringToSymEntry(strings, st, uint, errors), MsgType.NORMAL);
               }
               when "float64" {
-                  return new MsgTuple(castStringToSymEntry(strings, st, real), MsgType.NORMAL);
+                  return new MsgTuple(castStringToSymEntry(strings, st, real, errors), MsgType.NORMAL);
               }
               when "bool" {
-                  return new MsgTuple(castStringToSymEntry(strings, st, bool), MsgType.NORMAL);
+                  return new MsgTuple(castStringToSymEntry(strings, st, bool, errors), MsgType.NORMAL);
               }
               otherwise {
                  var errorMsg = notImplementedError(pn,"str",":",targetDtype);
@@ -153,99 +154,6 @@ module CastMsg {
         return new MsgTuple(errorMsg, MsgType.ERROR);
       }
     }
-  }
-
-  proc castGenSymEntry(gse: borrowed GenSymEntry, st: borrowed SymTab, type fromType, 
-                                             type toType): string throws {
-    const before = toSymEntry(gse, fromType);
-    const name = st.nextName();
-    var after = st.addEntry(name, before.size, toType);
-    try {
-      after.a = before.a : toType;
-    } catch e: IllegalArgumentError {
-      var errorMsg = "bad value in cast from %s to %s".format(fromType:string, 
-                                                       toType:string);
-      castLogger.error(getModuleName(),getRoutineName(),getLineNumber(),errorMsg);   
-      return "Error: %s".format(errorMsg);
-    }
-
-    var returnMsg = "created " + st.attrib(name);
-    castLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),returnMsg);
-    return returnMsg;
-  }
-
-  proc castGenSymEntryToString(gse: borrowed GenSymEntry, st: borrowed SymTab, 
-                                                       type fromType): string throws {
-    const before = toSymEntry(gse, fromType);
-    const oname = st.nextName();
-    var segments = st.addEntry(oname, before.size, int);
-    var strings: [before.aD] string;
-    if fromType == real {
-      try {
-          forall (s, v) in zip(strings, before.a) {
-              s = "%.17r".format(v);
-          }
-      } catch e {
-          var errorMsg = "could not convert float64 value to decimal representation";
-          castLogger.error(getModuleName(),getRoutineName(),getLineNumber(),errorMsg);   
-          return "Error: %s".format(errorMsg);
-      }
-    } else {
-      try {
-          strings = [s in before.a] s : string;
-      } catch e: IllegalArgumentError {
-          var errorMsg = "bad value in cast from %s to string".format(fromType:string);
-          castLogger.error(getModuleName(),getRoutineName(),getLineNumber(),errorMsg);   
-          return "Error: %s".format(errorMsg);
-      }
-    }
-    const byteLengths = [s in strings] s.numBytes + 1;
-    // check there's enough room to create a copy for scan and throw if creating a copy would go over memory limit
-    overMemLimit(numBytes(uint(8)) * byteLengths.size);
-    segments.a = (+ scan byteLengths) - byteLengths;
-    const totBytes = + reduce byteLengths;
-    const vname = st.nextName();
-    var values = st.addEntry(vname, totBytes, uint(8));
-    ref va = values.a;
-    forall (o, s) in zip(segments.a, strings) with (var agg = newDstAggregator(uint(8))) {
-      for (i, b) in zip(0.., s.bytes()) {
-        agg.copy(va[o+i], b);
-      }
-    }
-
-    var returnMsg ="created " + st.attrib(oname) + "+created " + st.attrib(vname);
-    castLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),returnMsg);
-    return returnMsg;
-  }
-
-  proc castStringToSymEntry(s: SegString, st: borrowed SymTab, type toType): string throws {
-      ref oa = s.offsets.a;
-      ref va = s.values.a;
-      const name = st.nextName();
-      var entry = st.addEntry(name, s.size, toType);
-    
-      const highInd = s.offsets.aD.high;
-      try {
-          forall (i, o, e) in zip(s.offsets.aD, s.offsets.a, entry.a) {
-              const start = o;
-              var end: int;
-
-              if (i == highInd) {
-              end = s.nBytes - 1;
-              } else {
-                   end = oa[i+1] - 1;
-              }
-              e = interpretAsString(va, start..end) : toType;
-          }
-      } catch e: IllegalArgumentError {
-          var errorMsg = "bad value in cast from string to %s".format(toType:string);
-          castLogger.error(getModuleName(),getRoutineName(),getLineNumber(),errorMsg);   
-          return "Error: %s".format(errorMsg);
-      }
-
-      var returnMsg = "created " + st.attrib(name);
-      castLogger.debug(getModuleName(),getRoutineName(),getLineNumber(),returnMsg);
-      return returnMsg;
   }
 
   proc registerMe() {