--- a/src/SegmentedComputation.chpl
+++ b/src/SegmentedComputation.chpl
@@ -1,6 +1,8 @@
 module SegmentedComputation {
   use CommAggregation;
   use SipHash;
+  use ServerErrors;
+  private use Cast;
 
   proc computeSegmentOwnership(segments: [?D] int, vD) {
     const low = vD.low;
@@ -36,9 +38,13 @@ module SegmentedComputation {
 
   enum SegFunction {
     SipHash128,
+    StringToNumericStrict,
+    StringToNumericIgnore,
+    StringToNumericReturnValidity,
   }
   
   proc computeOnSegments(segments: [?D] int, values: [?vD] ?t, param function: SegFunction, type retType) throws {
+    // type retType = if (function == SegFunction.StringToNumericReturnValidity) then (outType, bool) else outType;
     var res: [D] retType;
     if (D.size == 0) {
       return res;
@@ -59,16 +65,33 @@ module SegmentedComputation {
           agg.copy(mySegs[i], segments[i]);
           agg.copy(myLens[i], lengths[i]);
         }
-        // Apply function to bytes of each owned segment, aggregating return value to res
-        forall (start, len, i) in zip(mySegs, myLens, mySegInds) with (var agg = newDstAggregator(retType)) {
-          select function {
-            when SegFunction.SipHash128 {
-              agg.copy(res[i], sipHash128(values, start..#len));
-            }
-            otherwise {
-              compilerError("Unrecognized segmented function");
+        try {
+          // Apply function to bytes of each owned segment, aggregating return value to res
+          forall (start, len, i) in zip(mySegs, myLens, mySegInds) with (var agg = newDstAggregator(retType)) {
+            select function {
+              when SegFunction.SipHash128 {
+                agg.copy(res[i], sipHash128(values, start..#len));
+              }
+              when SegFunction.StringToNumericStrict {
+                agg.copy(res[i], stringToNumericStrict(values, start..#len, retType));
+              }
+              when SegFunction.StringToNumericIgnore {
+                agg.copy(res[i], stringToNumericIgnore(values, start..#len, retType));
+              }
+              when SegFunction.StringToNumericReturnValidity {
+                agg.copy(res[i], stringToNumericReturnValidity(values, start..#len, retType[0]));
+              }
+              otherwise {
+                compilerError("Unrecognized segmented function");
+              }
             }
           }
+        } catch {
+          throw new owned ErrorWithContext("Error computing %s on string or segment".format(function:string),
+                                           getLineNumber(),
+                                           getRoutineName(),
+                                           getModuleName(),
+                                           "IllegalArgumentError");
         }
       }
     }