This is part of #1101 and should improve performance of casting string arrays to numeric values when the locality of the string bytes is poor.

Performance is tracked in one of the benchmarks added in #1106 , but I have not yet been able to run this at a significant scale to verify the expected improvement.

While here, also add more robust error handling to str-to-num casts. It now supports three modes:
- strict (default): raise an exception if any value cannot be converted
- ignore: never raise an exception, bad inputs result in a dtype-dependent default/nan value
- return_validity: same result as 'ignore', but also return a bool array showing where the conversion succeeded.

I am not wild about adding another `Enum` for the error handling mode, because it feels rather annoying as a user. Can anyone think of a better way?