@ronawho 2022-02-15
16-node-cs-hdr performance:

Before:
```
Casting good locality Average time = 22.1167 sec
Casting good locality Average rate = 0.0758 GiB/sec
Casting poor locality Average time = 45.3778 sec
Casting poor locality Average rate = 0.0369 GiB/sec
```


Now:
```
Casting good locality Average time = 0.0950 sec
Casting good locality Average rate = 17.6440 GiB/sec
Casting poor locality Average time = 0.2211 sec
Casting poor locality Average rate = 7.5834 GiB/sec
```

In table form:

| config        | Before     | Now        |
| ------------- | ---------: | ---------: |
| good locality | 0.07 GiB/s | 17.6 GiB/s |
| poor locality | 0.03 GiB/s |  7.5 GiB/s |

So great improvements for good and poor locality.