@reuster986 2022-02-15
Wow, cool! I will have to go back and rationalize that speedup to myself, because I didn't expect it, but will definitely take it :)

Thanks again for running the measurement!