@pierce314159 2022-02-15
The only idea I have to avoid the `Enum` would be to have a separate cast function to handle the `ignore` and `return_validity` cases. So `cast` would handle the strict case and `relaxed_cast(return_validity = False)` would handle the other two. It's hard to tell if this would be more or less annoying though